fix(demo): menuHeaderRender - move check permission -
