+++ package.json
@@ -11,7 +11,7 @@
     "lint:nofix": "vue-cli-service lint --no-fix"
   },
   "dependencies": {
-    "@ant-design-vue/pro-layout": "^0.3.12",
+    "@ant-design-vue/pro-layout": "^1.0.1",
     "@antv/data-set": "^0.10.2",
     "ant-design-vue": "^1.6.2",
     "axios": "^0.19.0",
