+++ src/main.js
@@ -16,16 +16,18 @@ import themePluginConfig from '../config/themePluginConfig'
 import './mock'
 
 import bootstrap from './core/bootstrap'
-import './core/lazy_use'
+import './core/lazy_use' // use lazy load components
 import './permission' // permission control
 import './utils/filter' // global filter
-import './global.less'
+import './global.less' // global style
 
 Vue.config.productionTip = false
 
 // mount axios to `Vue.$http` and `this.$http`
 Vue.use(VueAxios)
+// use pro-layout components
 Vue.component('pro-layout', ProLayout)
+Vue.component('page-container', PageHeaderWrapper)
 Vue.component('page-header-wrapper', PageHeaderWrapper)
 
 window.umi_plugin_ant_themeVar = themePluginConfig.theme
@@ -34,6 +36,7 @@ new Vue({
   router,
   store,
   i18n,
+  // init localstorage, vuex
   created: bootstrap,
   render: h => h(App)
 }).$mount('#app')
