--- src/utils/permissions.js
@@ -1,8 +1,0 @@
-export function actionToObject (json) {
-  try {
-    return JSON.parse(json)
-  } catch (e) {
-    console.log('err', e.message)
-  }
-  return []
-}
