+++ src/views/other/modules/RoleModal.vue
@@ -23,7 +23,6 @@
 
 <script>
 import { getPermissions } from '@/api/manage'
-import { actionToObject } from '@/utils/permissions'
 import pick from 'lodash.pick'
 
 export default {
@@ -123,7 +122,7 @@ export default {
       getPermissions().then(res => {
         const result = res.result
         that.permissions = result.map(permission => {
-          const options = actionToObject(permission.actionData)
+          const options = JSON.parse(permission.actionData) || []
           permission.checkedAll = false
           permission.selected = []
           permission.indeterminate = false
